requirements.txt
Pillow
win32gui
